cron-service/requirements.txt
python-dotenv==0.19.1
requests==2.28.2
pybreaker==1.0.1
schedule==1.1.0